requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
idna==2.10
oauthlib==3.1.1
PySocks==1.7.1
python-dotenv==0.18.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
tinydb==4.5.0
tinydb-serialization==2.1.0
tweepy==3.10.0
urllib3==1.26.6
